Stabilizing Changes
Among the things I changed was the location the program expects to load
its settings file. It used to assume the file was in the same directory
as the binary, but this had some permission clashes with the default
user for apache. In order to make the program functional in the server
environment, I changed the path to /etc/html2faxcover/settings. Of
course, the program will fail if this file is not present. For
installation purposes, I made some fundamental changes to the Makefile.
The Makefile now has variables with the path for /usr/local/lib,
/etc/html2faxcover/, and /var/spool/html2faxcover/. With the exception
of the settings file, all other paths are changeable. By default, the
Makefile installs the binary in /var/spool/html2faxcover/ because I
found that to be a hassle free location when it comes to permissions.
However, by moving the settings file to the /etc/ directory, I think it
is possible to install the binary anywhere (keep in mind that you have
to modify the settings to match the changes). The other change was the
addition of the LGPL header on each file that did not have it previously
and I added some rules to the Makefile to make updating the settings
file easier! These are the main changes to the current stable release.