Major bugfixes
This commit contains changes that allow faxrecvd to process the
commandline options from the hylafax server. The file settings was also
expanded to include the faxrecvd settings with the html2faxcover
settings. Both programs in this project are now fully stable. I also
added a way to check for library version for the internal API.